refactored the blueprint parameters classes into a dynamic loadable module. preparing for changes that may occur after AOS 1.1
